+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## **Unreleased**
 **Hearthstone**
 - Improve secret tracker with secrets created by Mage's Quest reward The Origin Stone.
+- Added a counter for Infest the Scullery.
 - Fix Arenasmith tiles positioning in hover and redraft.
 
 ## **Release v1.53.10 - 2026-07-18**